LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-form-element/form-requestsubmit-autofocus.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<div id="target"></div>
<script>

async_test(t => {
  target.innerHTML = '<form><input id="submitter" type="submit"></input></form>';
  let form = document.querySelector('form');
  let submitter = document.querySelector('input');
  let iframe = document.createElement('iframe');
  iframe.onload = t.step_func_done(() => {
    assert_throws_js(TypeError, () => {
      form.requestSubmit(submitter);
    });
  });

  submitter.autofocus = true;
  submitter.onfocus = t.step_func(() => {
    submitter.type = 'button';
  });
  target.appendChild(iframe);
}, 'Due to autofocus an event handler can run that changes submitter type, so ' +
   'verify that it is processed before the requestSubmit checks.');

async_test(t => {
  target.innerHTML = '<form><input id="submitter" type="submit"></input></form>';
  let form = document.querySelector('form');
  let submitter = document.querySelector('input');
  let iframe = document.createElement('iframe');
  iframe.onload = t.step_func_done(() => {
    assert_throws_dom("NotFoundError", () => {
      form.requestSubmit(submitter);
    });
  });

  submitter.autofocus = true;
  submitter.onfocus = t.step_func(() => {
    form.removeChild(submitter);
  });
  target.appendChild(iframe);
}, 'Due to autofocus an event handler can run that removes the submitter, so ' +
   'verify that it is processed before the requestSubmit checks.');
</script>
</body>
